Fix size of uint32 in bittype

+++ Common/Libraries/Source/WWVegas/WWLib/bittype.h
@@ -37,20 +37,20 @@
  * - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - */
 #pragma once
 
+#include <cstdint>
+
 typedef unsigned char uint8;
 typedef unsigned short uint16;
-typedef unsigned long uint32;
+typedef uint32_t uint32;
 typedef unsigned int uint;
 
 typedef signed char sint8;
 typedef signed short sint16;
-typedef signed long sint32;
+typedef int32_t sint32;
 typedef signed int sint;
 
 typedef float float32;
 typedef double float64;
-
-#include <cstdint>
 
 typedef uint32_t DWORD;
 typedef uint16_t WORD;
